+++ modules/core/src/main/scala/moe/brianhsu/live2d/boundary/gateway/openSeeFace/OpenSeeFaceDataReader.scala
@@ -1,0 +1,9 @@
+package moe.brianhsu.live2d.boundary.gateway.openSeeFace
+
+import moe.brianhsu.live2d.enitiy.openSeeFace.OpenSeeFaceData
+
+trait OpenSeeFaceDataReader extends AutoCloseable {
+  def open(): Unit
+  def readData(): OpenSeeFaceData
+  def close(): Unit
+}