+++ CHANGELOG.md
@@ -4,5 +4,9 @@ CHANGELOG
 1.0.0
 ---------
 
+### Core API
+
+- Added Features
+  - OpenSeeFace UDP data reader API.
 - Breaking API Changes
   - Remove the adapter dependency in AvatarMotion factory method.